18/10/2023 - Game First Update
In this commit I have made scripts for movement, camera and a grapple movement. I have also created a weapon that will be used for the game as well as a material that will be seen throughout the game as the energy currency/economy. (had to recreate repo due to incorrect folder designations)